Camera coupled with entity system setup.

+++ Engine/Content/ContentToEngine.cpp
@@ -98,14 +98,14 @@ create_mesh_hierarchy(const void* const data) {
 	assert(lod_count);
 
 	geometry_hierarchy_stream stream{ hierarchy_buffer, lod_count };
-	u16 submesh_index{ 0 };
+	u32 submesh_index{ 0 };
 	id::id_type* const gpu_ids{ stream.gpu_ids() };
 
 	for (u32 lod_idx{ 0 }; lod_idx < lod_count; ++lod_idx) {
 		stream.thresholds()[lod_idx] = blob.read<f32>();
 		const u32 id_count{ blob.read<u32>() };
 		assert(id_count < (1 << 16));
-		stream.lod_offsets()[lod_idx] = { submesh_index, (u16)id_count };
+		stream.lod_offsets()[lod_idx] = { (u16)submesh_index, (u16)id_count };
 		blob.skip(sizeof(u32)); // skip over the size_of_submeshes
 		for (u32 id_idx{ 0 }; id_idx < id_count; ++id_idx) {
 			const u8* at{ blob.position() };
